test(bundler-core): relocate handleDevToolsRequest routing tests from devtools-vite
